Move things out of editor
