+++ packages/tldraw/src/lib/ui/components/EmbedDialog.tsx
@@ -1,11 +1,6 @@
-import {
-	EmbedDefinition,
-	isCustomEmbedDefinition,
-	isDefaultEmbedDefintionType,
-	track,
-	useEditor,
-} from '@tldraw/editor'
+import { TLEmbedDefinition, track, useEditor } from '@tldraw/editor'
 import { useRef, useState } from 'react'
+import { isCustomEmbedDefinition, isDefaultEmbedDefintionType } from '../../defaultEmbedDefinitions'
 import { EmbedShapeUtil } from '../../shapes/embed/EmbedShapeUtil'
 import { TLEmbedResult, getEmbedInfo } from '../../utils/embeds/embeds'
 import { useAssetUrls } from '../context/asset-urls'
@@ -29,7 +24,7 @@ export const EmbedDialog = track(function EmbedDialog({ onClose }: TLUiDialogPro
 	const assetUrls = useAssetUrls()
 
 	// The embed definition for the user's selected embed (set by the user clicking on an embed in stage 1)
-	const [embedDefinition, setEmbedDefinition] = useState<null | EmbedDefinition>(null)
+	const [embedDefinition, setEmbedDefinition] = useState<null | TLEmbedDefinition>(null)
 
 	// The URL that the user has typed into (in stage 2)
 	const [url, setUrl] = useState<string>('')
